Fix sorting order
It is not possible to order by date given the current storage format
(text formatted by PHP's date("F j, Y - H:i") function)

This patch is a first fix for existing installations to order snippets
by ID (which is autoincrementing).

Another patch that will follow aim to additionnaly store the date in
an ISO 8601 format.